+++ iosApp/iosApp/UI/Screens/ProjectList/ProjectInfo.swift
@@ -9,12 +9,12 @@
 import shared
 
 class ProjectInfo: TaskInfoProtocol {
-    var id: UInt8 = 0
+    var id: UInt16 = 0
     var title: String = "Нет названия"
     var timerValue: UInt8 = 0
     var participiantsValue: UInt8 = 0
 
-    init(id: UInt8 = .random(in: 1...99), title: String, timerValue: UInt8, participiantsValue: UInt8) {
+    init(id: UInt16 = .random(in: 1...99), title: String, timerValue: UInt8, participiantsValue: UInt8) {
         self.id = id
         self.title = title
         self.timerValue = timerValue
@@ -22,7 +22,7 @@ class ProjectInfo: TaskInfoProtocol {
     }
 
     init(_ dto: TaskDTO) {
-        id = dto.id as? UInt8 ?? self.id
+        id = dto.id as? UInt16 ?? self.id
         title = dto.name ?? self.title
         timerValue = dto.scope as? UInt8 ?? self.timerValue
         participiantsValue = dto.userCount as? UInt8 ?? self.timerValue
